requirements.txt
aws-cdk-lib>=2.220.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
psycopg2-binary>=2.9.0
